build: clear assorted -Werror warnings
- Texture2D.mm: cast the CGImageAlphaInfo|CGImageByteOrder ORs to uint32_t so the
  cross-enum bitwise operation is explicit.
- SVSegmentedThumb: mark the shadowColor/shadowOffset/castsShadow aliases @dynamic
  (their setters forward to the text* setters; no backing ivar).
- JcfUploader imports UIKit for its UIAlertView ivar.
- PurchaseManager reads +[JubeatAppDelegate clientInfo] as the class property it
  is, not through the appDelegate instance.
- Define the MainGamePhoneKntBgEffectSpec struct the knit background schedule uses.
- SVSegmentedControl casts the NSInteger index / int hoverIndex to NSUInteger in
  the two unsigned comparisons.
- Cast the NSInteger page / indexPath.row to int for the %02d howto/row image
  format strings.

+++ Project/ChallengeMenuView.m
@@ -402,7 +402,7 @@ static const UIViewAnimationOptions kVerifyFadeOptions = UIViewAnimationOptionCu
     cell.tag = indexPath.row;
     cell.aDelegate = self;
     UIImage *rowImage =
-        LoadScaledPngImage([NSString stringWithFormat:kRowImageFormat, indexPath.row]);
+        LoadScaledPngImage([NSString stringWithFormat:kRowImageFormat, (int)indexPath.row]);
     [cell setBgImage:rowImage numImage:numImageArray];
     if (indexPath.row == 0) {
         [cell setNumber:[ChallengeStatus sharedStatus].presentNum];
